// management/frontend/src/api/schemas/productKitPublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */
import type { KitDeviceSlot } from './kitDeviceSlot';

export interface ProductKitPublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 64
     */
  code: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  name: string;
  /** @maxLength 64 */
  product_type?: string;
  /** @maxLength 8000 */
  instructions?: string;
  exam_enabled?: boolean;
  /** @maxLength 32 */
  status?: string;
  sort?: number;
  device_slots?: KitDeviceSlot[];
  workspace_id: string;
  creator_name?: string | null;
  updater_name?: string | null;
}